[FLINK-23978][scala] Move Scala-reliant FieldAccessors to flink-streaming-scala
