+++ flink-streaming-java/src/main/java/org/apache/flink/streaming/util/typeutils/FieldAccessor.java
@@ -18,15 +18,12 @@
 package org.apache.flink.streaming.util.typeutils;
 
 import org.apache.flink.annotation.Internal;
-import org.apache.flink.api.common.ExecutionConfig;
 import org.apache.flink.api.common.typeinfo.BasicTypeInfo;
 import org.apache.flink.api.common.typeinfo.TypeInformation;
 import org.apache.flink.api.common.typeutils.CompositeType;
 import org.apache.flink.api.java.tuple.Tuple;
 import org.apache.flink.api.java.typeutils.TupleTypeInfo;
-import org.apache.flink.api.java.typeutils.TupleTypeInfoBase;
 import org.apache.flink.api.java.typeutils.runtime.FieldSerializer;
-import org.apache.flink.api.java.typeutils.runtime.TupleSerializerBase;
 
 import java.io.IOException;
 import java.io.ObjectInputStream;
@@ -34,8 +31,6 @@ import java.io.ObjectOutputStream;
 import java.io.Serializable;
 import java.lang.reflect.Array;
 import java.lang.reflect.Field;
-
-import scala.Product;
 
 import static org.apache.flink.util.Preconditions.checkNotNull;
 
@@ -299,109 +294,6 @@ public abstract class FieldAccessor<T, F> implements Serializable {
         private void readObject(ObjectInputStream in) throws IOException, ClassNotFoundException {
             in.defaultReadObject();
             field = FieldSerializer.deserializeField(in);
-        }
-    }
-
-    /**
-     * There are two versions of ProductFieldAccessor, differing in whether there is an other
-     * FieldAccessor nested inside. The no inner accessor version is probably a little faster.
-     */
-    static final class SimpleProductFieldAccessor<T, F> extends FieldAccessor<T, F> {
-
-        private static final long serialVersionUID = 1L;
-
-        private final int pos;
-        private final TupleSerializerBase<T> serializer;
-        private final Object[] fields;
-        private final int length;
-
-        SimpleProductFieldAccessor(int pos, TypeInformation<T> typeInfo, ExecutionConfig config) {
-            checkNotNull(typeInfo, "typeInfo must not be null.");
-            int arity = ((TupleTypeInfoBase) typeInfo).getArity();
-            if (pos < 0 || pos >= arity) {
-                throw new CompositeType.InvalidFieldReferenceException(
-                        "Tried to select "
-                                + ((Integer) pos).toString()
-                                + ". field on \""
-                                + typeInfo.toString()
-                                + "\", which is an invalid index.");
-            }
-
-            this.pos = pos;
-            this.fieldType = ((TupleTypeInfoBase<T>) typeInfo).getTypeAt(pos);
-            this.serializer = (TupleSerializerBase<T>) typeInfo.createSerializer(config);
-            this.length = this.serializer.getArity();
-            this.fields = new Object[this.length];
-        }
-
-        @SuppressWarnings("unchecked")
-        @Override
-        public F get(T record) {
-            Product prod = (Product) record;
-            return (F) prod.productElement(pos);
-        }
-
-        @Override
-        public T set(T record, F fieldValue) {
-            Product prod = (Product) record;
-            for (int i = 0; i < length; i++) {
-                fields[i] = prod.productElement(i);
-            }
-            fields[pos] = fieldValue;
-            return serializer.createInstance(fields);
-        }
-    }
-
-    static final class RecursiveProductFieldAccessor<T, R, F> extends FieldAccessor<T, F> {
-
-        private static final long serialVersionUID = 1L;
-
-        private final int pos;
-        private final TupleSerializerBase<T> serializer;
-        private final Object[] fields;
-        private final int length;
-        private final FieldAccessor<R, F> innerAccessor;
-
-        RecursiveProductFieldAccessor(
-                int pos,
-                TypeInformation<T> typeInfo,
-                FieldAccessor<R, F> innerAccessor,
-                ExecutionConfig config) {
-            int arity = ((TupleTypeInfoBase) typeInfo).getArity();
-            if (pos < 0 || pos >= arity) {
-                throw new CompositeType.InvalidFieldReferenceException(
-                        "Tried to select "
-                                + ((Integer) pos).toString()
-                                + ". field on \""
-                                + typeInfo.toString()
-                                + "\", which is an invalid index.");
-            }
-            checkNotNull(typeInfo, "typeInfo must not be null.");
-            checkNotNull(innerAccessor, "innerAccessor must not be null.");
-
-            this.pos = pos;
-            this.serializer = (TupleSerializerBase<T>) typeInfo.createSerializer(config);
-            this.length = this.serializer.getArity();
-            this.fields = new Object[this.length];
-            this.innerAccessor = innerAccessor;
-            this.fieldType = innerAccessor.getFieldType();
-        }
-
-        @SuppressWarnings("unchecked")
-        @Override
-        public F get(T record) {
-            return innerAccessor.get((R) ((Product) record).productElement(pos));
-        }
-
-        @SuppressWarnings("unchecked")
-        @Override
-        public T set(T record, F fieldValue) {
-            Product prod = (Product) record;
-            for (int i = 0; i < length; i++) {
-                fields[i] = prod.productElement(i);
-            }
-            fields[pos] = innerAccessor.set((R) fields[pos], fieldValue);
-            return serializer.createInstance(fields);
         }
     }
 }
