+++ flink-streaming-scala/src/test/java/org/apache/flink/streaming/util/typeutils/DefaultScalaProductFieldAccessorFactoryTest.java
@@ -1,0 +1,38 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.flink.streaming.util.typeutils;
+
+import org.apache.flink.util.TestLogger;
+
+import org.junit.Test;
+
+import java.util.Optional;
+
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.core.Is.is;
+
+/** Tests for the {@link DefaultScalaProductFieldAccessorFactory}. */
+public class DefaultScalaProductFieldAccessorFactoryTest extends TestLogger {
+    @Test
+    public void testLoad() {
+        final Optional<ScalaProductFieldAccessorFactory> load =
+                ScalaProductFieldAccessorFactory.load(log);
+
+        assertThat(load.isPresent(), is(true));
+    }
+}
